+++ src/helpers/balance/transferBalance.ts
@@ -3,7 +3,7 @@ import { SubmittableExtrinsicFunction } from '@polkadot/api/types'
 import { AccountId, EventRecord } from '@polkadot/types/interfaces'
 import { AnyTuple, Callback, IKeyringPair, ISubmittableResult } from '@polkadot/types/types'
 import { BN, bnToBn } from '@polkadot/util'
-import { ExstrinsicThrowErrorMessage, getExtrinsicErrorMessage } from '../getExtrinsicErrorMessage'
+import { ExstrinsicThrowErrorMessage, getExtrinsicErrorMessage } from '../contracts/getExtrinsicErrorMessage'
 import { checkIfBalanceSufficient } from './checkIfBalanceSufficient'
 
 export type TransferBalanceResult = {
