+++ src/helpers/contracts/contractCall.ts
@@ -5,10 +5,10 @@ import { EventRecord } from '@polkadot/types/interfaces'
 import { Callback, IKeyringPair, ISubmittableResult } from '@polkadot/types/types'
 import { BN, stringCamelCase } from '@polkadot/util'
 import { checkIfBalanceSufficient } from '../balance/checkIfBalanceSufficient'
-import { decodeOutput } from '../decodeOutput'
-import { getExtrinsicErrorMessage } from '../getExtrinsicErrorMessage'
-import { getMaxGasLimit } from '../getGasLimit'
+import { decodeOutput } from './decodeOutput'
 import { getAbiMessage } from './getAbiMessage'
+import { getExtrinsicErrorMessage } from './getExtrinsicErrorMessage'
+import { getMaxGasLimit } from './getGasLimit'
 
 /**
  * Performs a dry run for the given contract method and arguments.
