+++ src/helpers/index.ts
@@ -1,17 +1,17 @@
 // @index('./*.(tsx|ts)', f => `export * from '${f.path}'`)
-export * from './accountsAreEqual'
+export * from './balance/accountsAreEqual'
 export * from './balance/checkIfBalanceSufficient'
 export * from './balance/formatBalance'
 export * from './balance/getBalance'
 export * from './balance/getPSP22Balances'
 export * from './balance/transferBalance'
 export * from './contracts/contractCall'
+export * from './contracts/decodeOutput'
 export * from './contracts/deployContract'
 export * from './contracts/getAbi'
 export * from './contracts/getAbiMessage'
-export * from './decodeOutput'
-export * from './getExtrinsicErrorMessage'
-export * from './getGasLimit'
+export * from './contracts/getExtrinsicErrorMessage'
+export * from './contracts/getGasLimit'
+export * from './contracts/unwrapResult'
 export * from './initPolkadotJs'
-export * from './unwrapResult'
 // @endindex
